Convert board into a function component
