--- lib/chess_web/board_live.html.heex
@@ -1,23 +1,0 @@
-<section>
-  <div class="board">
-    <%= for y <- 8..1 do %>
-      <%= for x <- 1..8 do %>
-        <div class={"square #{get_square_color(x, y, @selected_square)}"} id={"#{x}#{y}"}
-          phx-click="move_piece_to"
-          phx-target={@myself}
-          phx-value-x={x}
-          phx-value-y={y}
-        >
-          <%= if is_piece(x, y, @pieces) do %>
-            <div class={"figure #{is_piece(x, y, @pieces)}"} 
-              phx-click="move_piece_from"
-              phx-target={@myself}
-              phx-value-x={x}
-              phx-value-y={y}
-        />
-          <% end %>  
-        </div>
-      <% end %>
-    <% end %>
-  </div>
-</section>
