+++ lib/chess_web/game_live.html.heex
@@ -1,5 +1,4 @@
-<.live_component
-  module={ChessWeb.BoardLive}
+<.board
   id={@chess_game_topic_id}
   pieces={@pieces}
   selected_square={@selected_square} />
